frontend/requirements.txt
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.0
Pillow>=9.0.0
requests>=2.28.0
numpy>=1.21.0
openai>=1.0.0
